Provide a direct artifact containing 14.38.33130
Per request from some users, I'm providing the correct MSVC compiler to use with our version pinning, as it can be non-trivial to acquire and our goal is an instantly reproducible build. This does add quite a bit to the overall weight of barrage, which is frustrating and unfortunate, but there's actually not a particularly good way to hardlink to these files.