Fixes to prevent type leakage.
Probably the ugliest fix I've used in ages. I wouldn't normally do something like this, but I'm tired, in a hurry, and I want to be absolutely sure it'll be insanely fast. Placement new is basically made specifically for this, and with the assert in there, it's pretty safe.

This isn't a cast, let's be clear. it probably should be, but there's a bunch of issues with how to structure and forward declare that which I'm actually pretty uncomfortable with and would defeat most of the advantages of the type encapsulation we've been working towards.

I'll revisit this later, as it has some dangerous failure modes and unnecessary runtime costs.