Auto is contraindicated for anything other than known types.
To be honest, I think I may actually add a lint rule for this. Ultimately, it was auto that got us here combined with the anomalous behavior of the stupid bloody TType generics from UE and MakeShareable. Basically, auto captured a raw pointer type internal to the UE shared pointer, not a shared pointer, for reasons I don't understand and don't really care about. I'd never have caught this except for rider, actually, so kudos to type inferring tools. MakeShareable _really_ shouldn't produce this type anyway. This is why we don't reimplement shared pointer, folks. Even the very best engineers struggled with it for years to get it really elegant.

So this actually would have been fine, we'd normally have just leaked memory. but fun story:
Your friend and mine, the TType generics? Yeah, it looks like they _delete anything stored in them_ when destroyed  AND this meant that we got a double free, since we DO actually get a shared pointer out of manage pointers by invisible type coercion due to the return type. So, when the ref count hits zero, we free the primitive, then when the container is destroyed, it freed the raw pointer.

To be honest, just raw-pointer frees are probably enough, but the end result of all of this is a double free.