math is now working
ratio's not correct, unfortunately, it looks like, or some aspect of delta t isn't being correctly applied but overall, this is a marked improvement. we're still not seeing movement, but oddly, the vector state of the effective velocity vector is updating. I wonder if the player kine is still not quite right?